+++ R/mortality.R
@@ -17,7 +17,7 @@ NULL
 #' mortality at location j within t or fewer steps if starting at location i.
 #'
 #' The returned matrix will always be dense and cannot be optimized. Must enable
-#' override to use.
+#' override to use (see \code{\link{samc-class}}).
 #'
 #'   \item \strong{mortality(samc, origin, time)}
 #'
@@ -63,7 +63,7 @@ NULL
 #' mortality at location j if starting at location i.
 #'
 #' The returned matrix will always be dense and cannot be optimized. Must enable
-#' override to use.
+#' override to use (see \code{\link{samc-class}}).
 #'
 #'   \item \strong{mortality(samc, origin)}
 #'
@@ -122,7 +122,7 @@ setMethod(
   signature(samc = "samc", occ = "missing", origin = "missing", dest = "missing", time = "numeric"),
   function(samc, time) {
     if (!samc@override)
-      stop("This version of the mortality() method produces a large dense matrix.\nIn order to run it, create the samc object with the override parameter set to TRUE.", call. = FALSE)
+      stop("This version of the mortality() method produces a large dense matrix.\nSee the documentation for details.", call. = FALSE)
 
     if (time %% 1 != 0 || time < 1 || length(time) > 1)
       stop("The time argument must be a single positive integer", call. = FALSE)
@@ -272,7 +272,7 @@ setMethod(
   signature(samc = "samc", occ = "missing", origin = "missing", dest = "missing", time = "missing"),
   function(samc) {
     if (!samc@override)
-      stop("This version of the mortality() method produces a large dense matrix.\nIn order to run it, create the samc object with the override parameter set to TRUE.", call. = FALSE)
+      stop("This version of the mortality() method produces a large dense matrix.\nSee the documentation for details.", call. = FALSE)
 
     f <- visitation(samc)
     gc()