+++ R/samc-class.R
@@ -13,16 +13,29 @@
 #' that users are less likely to inadvertently alter it in a way that will cause
 #' issues in calculations.
 #'
-#' The class also contains a RasterLayer object derived from the input data.
-#' This object is used for checking inputs and mapping vector data in other
-#' functions.
-#'
-#' Finally, an override flag is used to help ensure that users do not
-#' accidentally run memory intensive versions of functions that can cause their
-#' systems to become non-responsive or for software to crash.
-#'
-#' The \code{\link{samc}} function is used to create \code{\link{samc-class}}
+#' The \code{\link{samc()}} function is used to create \code{\link{samc-class}}
 #' objects.
+#'
+#' The samc-class slots are subject to change, so users should not be using the
+#' \code{@} operator to access or change them. Doing so leads to the risk of broken
+#' code in the future. Instead, where relevant, special functions are supplied to
+#' get and set the values in the slots safely. This is a current list:
+#'
+#' \itemize{
+#'   \item \code{override()}
+#'
+#'   Some analyses are memory intensive and have the potential to make a user's
+#'   system non-responsive or crash. By default, a samc-class object cannot be used
+#'   in these analyses to prevent unintentional loss of work. In some cases, users
+#'   may wish to use these particular analyses, in which case this behavior can
+#'   be overridden. To get the current state of the override, use \code{override(samc)}.
+#'   To enable the use of the analyses, the override can be set to \code{TRUE}, for example:
+#'   \code{override(samc) <- TRUE}. Before enabling the override, users should
+#'   familiarize themselves with the Performance vignette. They should also consider
+#'   starting with scaled down versions of their data and then gradually scale
+#'   back up while monitoring their memory usage to guage what their system can
+#'   handle.
+#' }
 #'
 #' @slot p The transition probability matrix \emph{P}.
 #' @slot source Information about the data source for the P matrix
@@ -51,3 +64,47 @@ setClass(
   #   return(TRUE)
   # }
   )
+
+setGeneric("override",
+           function(samc) {
+             standardGeneric("override")
+           })
+
+setMethod("override",
+          signature(samc = "samc"),
+          function(samc) {
+            return(samc@override)
+          })
+
+setGeneric("override<-",
+           function(samc, value) {
+             standardGeneric("override<-")
+           })
+
+setMethod("override<-",
+          signature(samc = "samc", value = "logical"),
+          function(samc, value) {
+            samc@override <- value
+            return(samc)
+          })
+
+
+
+# setMethod( "$<-", signature(x = "samc"), function(x, name, value) {
+#   if(name=="override"){
+#     x@override <- value
+#   } else {
+#     warning("Invalid option name specified.", call. = FALSE)
+#   }
+#   return(x)
+# })
+#
+#
+# setMethod( "$", signature(x = "samc"), function(x, name) {
+#   if(name=="override"){
+#     return(x@override)
+#   } else {
+#     warning("Invalid option name specified.", call. = FALSE)
+#   }
+#   return(NULL)
+# })