+++ NEWS.md
@@ -1,7 +1,8 @@
 # Dev
 
-- Added `symm` parameter to samc() for transition functions
-- Updated long/lat handling in samc(). Deprecated `latlon` parameter. Added warning for rasters with non-square cells and missing projection.
+- Added symm parameter to samc() for transition functions
+- Updated long/lat handling in samc(). Deprecated latlon parameter. Added warning for rasters with non-square cells and missing projection.
+- Deprecated override parameter in the samc() function. Added override() get/set methods for samc-class objects. See documentation for samc-class for details.
 
 # samc 1.3.0